# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API TICKETS",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000"
        }
    ],
    "paths": {
...
        "/api/items/{id}/orders/{buyer}": {
            "get": {
                "tags": [
                    "Orders"
                ],
                "summary": "select interested buyer",
                "description": "As a Seller, I want to select an interested Buyer and mark an item as “Sold” so that the Item will no longer be listed on the home page.",
                "operationId": "selectInterestedBuyer",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Item id",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    },
                    {
                        "name": "buyer",
                        "in": "path",
                        "description": "Buyer id",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "400": {
                        "description": "Invalid ID supplier"
                    },
                    "404": {
                        "description": "Item not found"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/api/items/{id}/order": {
            "post": {
                "tags": [
                    "Orders"
                ],
                "summary": "Order an item",
                "description": "As a Buyer, I want to perform an action to show my interest for an Item and enter my name, email and location so that the Seller can see my information and contact me later.",
                "operationId": "orderStore",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of Item to make Order",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "quantity",
                                    "country",
                                    "street",
                                    "state",
                                    "firstName",
                                    "email",
                                    "lastName",
                                    "local_government"
                                ],
                                "properties": {
                                    "firstName": {
                                        "type": "text"
                                    },
                                    "lastName": {
                                        "type": "text"
                                    },
                                    "email": {
                                        "type": "email"
                                    },
                                    "state": {
                                        "type": "street"
                                    },
                                    "street": {
                                        "type": "street"
                                    },
                                    "local_government": {
                                        "type": "text"
                                    },
                                    "country": {
                                        "type": "text"
                                    },
                                    "quantity": {
                                        "type": "text"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Success"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearer_token": {
                "type": "apiKey",
                "description": "Sanctum Authentication ",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}